:fire: chore(frontend): 删除事件系统 — events/ 整目录 + pipeline.rs 中 19+ 处 emit 调用
