+++ src/frontend/mod.rs
@@ -66,7 +66,6 @@ pub mod pipeline;
 pub use crate::util::diagnostic;
 
 // 事件系统
-pub mod events;
 pub mod validate;
 
 // 编译器核心（事件驱动）
@@ -88,7 +87,6 @@ pub use pipeline::{Pipeline, PipelineState, CompilationResult};
 pub use compiler::CompileError;
 
 // 事件类型
-pub use events::*;
 pub use validate::{validate_source, ValidateResult};
 
 #[cfg(test)]
