+++ src/middle/passes/mono/mod.rs
@@ -13,7 +13,7 @@ pub mod instance;
 pub mod type_mono;
 
 use function::FunctionMonomorphizer;
-use instance::{GenericFunctionId, InstantiationRequest, SpecializationKey, TypeId};
+use instance::{GenericFunctionId, InstantiationRequest, SpecializationKey};
 use crate::frontend::core::typecheck::MonoType;
 use crate::middle::core::ir::{BasicBlock, ConstValue, FunctionIR, Instruction, ModuleIR, Operand};
 
@@ -31,9 +31,6 @@ pub struct Monomorphizer {
     max_depth: usize,
     /// 泛型类型定义：type_name -> MonoType（含 TypeVar）
     generic_types: HashMap<String, MonoType>,
-    /// 已单态化的类型：TypeId -> MonoType
-    #[allow(dead_code)]
-    monomorphized_types: HashMap<TypeId, MonoType>,
 }
 
 impl Monomorphizer {
@@ -45,7 +42,6 @@ impl Monomorphizer {
             processed: HashSet::new(),
             max_depth: 100,
             generic_types: HashMap::new(),
-            monomorphized_types: HashMap::new(),
         }
     }
